+++ forms/src/features/numberInput/directives/index.ts
@@ -1,0 +1,1 @@
+export * from './numberInputFormControl/numberInputFormControl.directive';